LayoutTests/inspector-obsolete/http-tests/network/network-size-expected-chunked.txt

CONSOLE MESSAGE: line 22: Done.
Tests network resource size reporting.


http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&size=500&gzip=true%chunked=true
resourceSize: 500
transferSize: 208
